+++ src/main/java/com/ifengxue/plugin/generator/merge/JavaSourceFileMerger.java
@@ -6,7 +6,6 @@ import com.ifengxue.plugin.Holder;
 import com.ifengxue.plugin.entity.Table;
 import com.ifengxue.plugin.generator.config.GeneratorConfig;
 import com.intellij.lang.jvm.annotation.JvmAnnotationAttribute;
-import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.psi.PsiAnnotation;
 import com.intellij.psi.PsiClass;
@@ -62,8 +61,8 @@ public class JavaSourceFileMerger implements SourceFileMerger {
                 PsiClass resolvePsiClass = implementsListType.resolve();
                 if (resolvePsiClass != null && !nameToClassType
                     .containsKey(resolvePsiClass.getQualifiedName())) {
-                    PsiJavaCodeReferenceElement implementsReference = ServiceManager
-                        .getService(Holder.getOrDefaultProject(), PsiElementFactory.class)
+                    PsiJavaCodeReferenceElement implementsReference = Holder.getOrDefaultProject()
+                        .getService(PsiElementFactory.class)
                         .createReferenceFromText(
                             Objects.requireNonNull(resolvePsiClass.getQualifiedName()),
                             originalTopClass);