+++ src/main/java/com/ifengxue/plugin/gui/property/JavaDataTypeEditorProvider.java
@@ -3,8 +3,8 @@ package com.ifengxue.plugin.gui.property;
 import com.ifengxue.plugin.Holder;
 import com.ifengxue.plugin.gui.annotation.EditorProvider;
 import com.ifengxue.plugin.util.TypeUtil;
-import com.intellij.openapi.editor.event.DocumentAdapter;
 import com.intellij.openapi.editor.event.DocumentEvent;
+import com.intellij.openapi.editor.event.DocumentListener;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectManager;
 import java.beans.PropertyDescriptor;
@@ -14,13 +14,14 @@ public class JavaDataTypeEditorProvider implements EditorProvider {
 
   @Override
   public TableCellEditor createEditor(PropertyDescriptor pd) {
+    @SuppressWarnings("deprecation")
     Project project = Holder.getProject();
     if (project == null) {
       project = ProjectManager.getInstance().getDefaultProject();
     }
     ClassNameAutoCompletionTableCellEditor editor = new ClassNameAutoCompletionTableCellEditor(project,
         TypeUtil.getAllJavaDbType());
-    editor.addDocumentListener(new DocumentAdapter() {
+    editor.addDocumentListener(new DocumentListener() {
       @Override
       public void documentChanged(DocumentEvent e) {